pyproject.toml
[tool.poetry]
name = "claude-agent-sdk-demo"
version = "0.1.0"
description = "Demonstration of Claude Agent SDK with security hazard illustrations"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
claude-agent-sdk = "*"
anthropic = "*"
anyio = "*"
python-dotenv = "*"
streamlit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
